third_party/WebKit/LayoutTests/accessibility/aria-label.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">
<input id="text" aria-label="aria label" type="text" size=20>
<a id="link" href="#" aria-label="aria link">test</a>
<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria-label attribute works. The input and the A tag should return the appropriate description.");

    if (window.accessibilityController) {

          var obj = accessibilityController.accessibleElementById("text");
          var succeeded = obj.name == "aria label";
          shouldBe("succeeded", "true");

          obj = accessibilityController.accessibleElementById("link");
          succeeded = obj.name == "aria link";
          shouldBe("succeeded", "true");
    }

</script>

</body>
</html>
